% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyclestreets_route}
\alias{cyclestreets_route}
\title{A journey from CycleStreets.net}
\format{
An object of class \code{sf}
}
\source{
CycleStreets.net
}
\usage{
cyclestreets_route
}
\description{
Road segments representing suggested route to cycle
in Leeds, UK.
}
\details{
Simple feature collection with 30 features and 32 fields

See \code{data-raw} folder in the package's github repo for details.
}
\examples{
library(sf)
class(cyclestreets_route)
plot(cyclestreets_route$geometry)
}
\keyword{datasets}
