% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dem_lisbon}
\alias{dem_lisbon}
\title{Read the bundled Lisbon DEM as a SpatRaster}
\usage{
dem_lisbon()
}
\value{
A \code{SpatRaster} object with 133 rows, 200 columns, and 1 elevation layer.
}
\description{
Returns a \code{SpatRaster} (terra package) of the Digital Elevation Model
for central Lisbon, Portugal, bundled with the slopes package.
}
\examples{
dem_lisbon()
}
