% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slope.R
\name{distance_z}
\alias{distance_z}
\title{Extract distance and elevation data from route}
\usage{
distance_z(route_xyz, lonlat)
}
\arguments{
\item{route_xyz}{An sf object with XYZ coordinates}

\item{lonlat}{Logical, whether coordinates are longitude/latitude}
}
\value{
List with components d (distances) and z (elevations)
}
\description{
Extracts cumulative distance and elevation vectors from route XYZ coordinates.
}
