% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{elevation_extract}
\alias{elevation_extract}
\title{Extract elevations from coordinates}
\usage{
elevation_extract(
  m,
  dem,
  method = "bilinear",
  terra = has_terra() && methods::is(dem, "SpatRaster")
)
}
\arguments{
\item{m}{Coordinates matrix or sf object}

\item{dem}{RasterLayer or SpatRaster}

\item{method}{character; extraction method (default "bilinear")}

\item{terra}{logical; use terra if TRUE}
}
\description{
Extract elevations from coordinates
}
