% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{elevation_get}
\alias{elevation_get}
\title{Get elevation data for routes}
\usage{
elevation_get(routes, ...)
}
\arguments{
\item{routes}{An sf object containing linestring geometries}

\item{...}{Additional arguments passed to ceramic::cc_elevation}
}
\value{
A SpatRaster covering the routes
}
\description{
Downloads elevation data using the ceramic package for given routes.
Returns a \code{SpatRaster} object (terra package).
}
