% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_get.R
\name{elevations_get}
\alias{elevations_get}
\title{Get elevation data from hosted maptile services}
\usage{
elevations_get(r, file = NULL, ...)
}
\arguments{
\item{r}{Routes, the gradients of which are to be calculated}

\item{file}{Where to save the resulting data if specified (not implemented)}
}
\description{
Get elevation data from hosted maptile services
}
\examples{
r = cyclestreets_route
e = elevations_get(r)
class(e)
e
plot(e)
plot(sf::st_geometry(r), add = TRUE)
}
