% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{has_terra}
\alias{has_terra}
\title{Check if terra package is available}
\usage{
has_terra()
}
\value{
Logical value indicating if terra is available
}
\description{
Checks whether the terra package is installed and can be loaded.
}
