% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lisbon_road_segment}
\alias{lisbon_road_segment}
\alias{lisbon_road_segment_3d}
\alias{lisbon_route}
\alias{lisbon_route_3d}
\title{A road segment in Lisbon, Portugal}
\format{
An object of class \code{sf}
}
\source{
Produced by ESRI's \href{https://pro.arcgis.com/en/pro-app/help/analysis/3d-analyst/get-started-with-3d-analyst-in-pro.htm}{3D Analyst extension}
}
\usage{
lisbon_road_segment
}
\description{
A single road segment and a 3d version.
}
\examples{
lisbon_road_segment
lisbon_road_segment_3d
}
\keyword{datasets}
