% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lisbon_road_segment_3d}
\alias{lisbon_road_segment_3d}
\title{Lisbon road segment 3D}
\format{
An sf object
}
\usage{
lisbon_road_segment_3d
}
\description{
Lisbon road segment 3D
}
\keyword{datasets}
