% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lisbon_road_segment_xyz_mapbox}
\alias{lisbon_road_segment_xyz_mapbox}
\title{Lisbon road segment XYZ}
\format{
An sf object
}
\usage{
lisbon_road_segment_xyz_mapbox
}
\description{
Lisbon road segment XYZ
}
\keyword{datasets}
