% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slope.R
\name{make_breaks}
\alias{make_breaks}
\title{Create slope breaks for color mapping}
\usage{
make_breaks(brks)
}
\arguments{
\item{brks}{Vector of positive slope break values (as percentages)}
}
\value{
Vector of slope breaks including negative values and zero
}
\description{
Creates symmetric slope breaks around zero for color classification.
}
