% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slope.R
\name{make_colz}
\alias{make_colz}
\title{Assign colors to slope values}
\usage{
make_colz(g, b, pal)
}
\arguments{
\item{g}{Vector of slope gradient values}

\item{b}{Vector of break points}

\item{pal}{Vector of colors corresponding to breaks}
}
\value{
Character vector of colors for each slope value
}
\description{
Maps slope gradient values to colors based on break points.
}
