% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{sequential_dist}
\alias{sequential_dist}
\title{Calculate sequential distances between coordinate pairs}
\usage{
sequential_dist(m, lonlat = TRUE)
}
\arguments{
\item{m}{Matrix of coordinates}

\item{lonlat}{Are coordinates lon/lat? TRUE by default}
}
\value{
Vector of distances
}
\description{
Calculate sequential distances between coordinate pairs
}
