% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_3d}
\alias{slope_3d}
\title{Take a linestring and add a third (z) dimension to its coordinates}
\usage{
slope_3d(r, e, method = "bilinear")
}
\arguments{
\item{r}{Routes, the gradients of which are to be calculated}

\item{e}{A raster object overlapping with \code{r} and values representing elevations}

\item{method}{The method of estimating elevation at points, passed to the \code{extract}
function for extracting values from raster datasets. Default: \code{"bilinear"}.}
}
\description{
Take a linestring and add a third (z) dimension to its coordinates
}
\examples{
r = lisbon_road_segments[239, ]
e = dem_lisbon_raster
(r3d = slope_3d(r, e))
sf::st_z_range(r)
sf::st_z_range(r3d)
}
