% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_distance}
\alias{slope_distance}
\title{Calculate slopes using distance data}
\usage{
slope_distance(d, elevations)
}
\arguments{
\item{d}{Vector of distance values between points}

\item{elevations}{Vector of elevation values}
}
\value{
Numeric vector of slope values
}
\description{
Calculates slope gradients from distance and elevation vectors.
}
