% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_distance_mean}
\alias{slope_distance_mean}
\title{Calculate mean slope using distance weighting}
\usage{
slope_distance_mean(d, elevations, directed = FALSE)
}
\arguments{
\item{d}{Vector of distance values between points}

\item{elevations}{Vector of elevation values}

\item{directed}{Logical, whether to calculate directed slopes (default: FALSE)}
}
\value{
Numeric value representing the mean slope
}
\description{
Computes the mean slope across segments using distance-weighted averaging.
}
