% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_distance_weighted}
\alias{slope_distance_weighted}
\title{Calculate distance-weighted slopes}
\usage{
slope_distance_weighted(d, elevations, directed = FALSE)
}
\arguments{
\item{d}{Vector of distance values between points}

\item{elevations}{Vector of elevation values}

\item{directed}{Logical, whether to calculate directed slopes (default: FALSE)}
}
\value{
Numeric value representing the weighted slope
}
\description{
Applies distance-based weighting to slope calculations for more accurate results.
}
