% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_matrices}
\alias{slope_matrices}
\title{Apply slope function over multiple matrices}
\usage{
slope_matrices(m_xyz_split, fun = slope_matrix_weighted, ...)
}
\arguments{
\item{m_xyz_split}{List of matrices/data.frames with xyz coordinates}

\item{fun}{Slope function to apply (default: slope_matrix_weighted)}

\item{...}{Additional arguments to fun}
}
\description{
Apply slope function over multiple matrices
}
