% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_get.R
\name{slope_matrix_to_raster}
\alias{slope_matrix_to_raster}
\title{Convert a slope matrix to a raster object}
\usage{
slope_matrix_to_raster(x)
}
\arguments{
\item{x}{Matrix or RasterLayer of slope values}
}
\value{
A RasterLayer object
}
\description{
Convert a slope matrix to a raster object
}
