% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_matrix_weighted}
\alias{slope_matrix_weighted}
\title{Calculate weighted slopes from coordinate matrix}
\usage{
slope_matrix_weighted(m, elevations = m[, 3], lonlat = TRUE, directed = FALSE)
}
\arguments{
\item{m}{Matrix of coordinates (x, y, z)}

\item{elevations}{Vector of elevation values (default: third column of m)}

\item{lonlat}{Logical, whether coordinates are longitude/latitude (default: TRUE)}

\item{directed}{Logical, whether to calculate directed slopes (default: FALSE)}
}
\value{
Numeric value representing the weighted slope
}
\description{
Applies distance-based weighting to slope calculations from coordinate matrix.
}
