% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_xyz}
\alias{slope_xyz}
\title{Calculate slopes from xyz data frames or sf objects}
\usage{
slope_xyz(
  route_xyz,
  fun = slope_matrix_weighted,
  lonlat = TRUE,
  directed = FALSE
)
}
\arguments{
\item{route_xyz}{data.frame or sf object}

\item{fun}{slope function (default: slope_matrix_weighted)}

\item{lonlat}{logical; are coordinates longitude/latitude?}

\item{directed}{logical; if TRUE, downslope is negative}
}
\description{
Calculate slopes from xyz data frames or sf objects
}
