% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_get.R
\name{slope_xyz_simple}
\alias{slope_xyz_simple}
\title{Extract XYZ coordinates from raster or matrix}
\usage{
slope_xyz_simple(x)
}
\arguments{
\item{x}{A RasterLayer or matrix object}
}
\value{
A data frame with x, y, z coordinates
}
\description{
Simplifies raster or matrix data to XYZ coordinate format.
}
