% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.R
\name{slope_z_value}
\alias{slope_z_value}
\alias{slope_z_start}
\alias{slope_z_end}
\alias{slope_z_mean}
\alias{slope_z_max}
\alias{slope_z_min}
\title{Calculate summary values for 'Z' elevation attributes}
\usage{
slope_z_value(x)

slope_z_start(x)

slope_z_end(x)

slope_z_mean(x)

slope_z_max(x)

slope_z_min(x)
}
\arguments{
\item{x}{An \code{sfc} object with 'XYZ' coordinates}
}
\description{
The \verb{slope_z*()} functions calculate summary values for the Z axis
in \code{sfc} objects with \code{XYZ} geometries.
}
\examples{
x = slopes::lisbon_route_3d
sf::st_geometry(x)
slope_z_value(x)[1:5]
xy = slopes::lisbon_route
# slope_z_value(xy) # error message
slope_z_start(x)
slope_z_end(x)
}
