% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes-package.R
\docType{package}
\name{slopes-package}
\alias{slopes}
\alias{slopes-package}
\title{slopes: Calculate Slopes of Roads, Rivers and Trajectories}
\value{
The package does not return a single value, but a collection of functions.
}
\description{
Calculate the slope (also known as steepness and gradient)
of linear features such as roads, rivers and ski routes.
}
\details{
Created by Robin Lovelace and Rosa Félix, who developed the package
to support their research in active travel, and the prioritisation of
investment in cycling, an activity which is sensitive to slopes,
in particular.

See \verb{https://github.com/ropensci/slopes/}
for the source code and \verb{https://docs.ropensci.org/slopes/}
for the website.

All package functions can be found in the
\href{https://docs.ropensci.org/slopes/reference/index.html}{Reference page}
on the package website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/slopes/}
  \item \url{https://docs.ropensci.org/slopes/}
  \item Report bugs at \url{https://github.com/ropensci/slopes/issues}
}

}
\author{
\strong{Maintainer}: Robin Lovelace \email{rob00x@gmail.com} (\href{https://orcid.org/0000-0001-5679-6536}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Rosa Félix \email{rosamfelix@tecnico.ulisboa.pt} (\href{https://orcid.org/0000-0002-5642-6006}{ORCID})
  \item Joey Talbot \email{j.d.talbot@leeds.ac.uk} (\href{https://orcid.org/0000-0002-6520-4560}{ORCID})
  \item Gonçalo Matos \email{goncaloafmatos@tecnico.ulisboa.pt} (\href{https://orcid.org/0009-0001-3489-1732}{ORCID})
}

Other contributors:
\itemize{
  \item Dan Olner (Dan reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/420#issuecomment-857662657) [reviewer]
  \item Andy Teucher (Andy reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/420#issuecomment-858231647) [reviewer]
}

}
\keyword{internal}
