% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-functions.R
\name{z_end}
\alias{z_end}
\title{Extract ending elevation from xyz linestring}
\usage{
z_end(x)
}
\arguments{
\item{x}{An sf object with XYZ coordinates}
}
\value{
A numeric value representing ending elevation
}
\description{
Extract ending elevation from xyz linestring
}
