% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-functions.R
\name{z_max}
\alias{z_max}
\title{Extract maximum elevation from xyz linestring}
\usage{
z_max(x)
}
\arguments{
\item{x}{An sf object with XYZ coordinates}
}
\value{
A numeric value representing maximum elevation
}
\description{
Extract maximum elevation from xyz linestring
}
