% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-functions.R
\name{z_mean}
\alias{z_mean}
\title{Extract mean elevation from xyz linestring}
\usage{
z_mean(x)
}
\arguments{
\item{x}{An sf object with XYZ coordinates}
}
\value{
A numeric value representing mean elevation
}
\description{
Extract mean elevation from xyz linestring
}
