% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-functions.R
\name{z_min}
\alias{z_min}
\title{Extract minimum elevation from xyz linestring}
\usage{
z_min(x)
}
\arguments{
\item{x}{An sf object with XYZ coordinates}
}
\value{
A numeric value representing minimum elevation
}
\description{
Extract minimum elevation from xyz linestring
}
