% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-functions.R
\name{z_start}
\alias{z_start}
\title{Extract starting elevation from xyz linestring}
\usage{
z_start(x)
}
\arguments{
\item{x}{An sf object with XYZ coordinates}
}
\value{
A numeric value representing starting elevation
}
\description{
Extract starting elevation from xyz linestring
}
