% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cyclestreets_route}
\alias{cyclestreets_route}
\title{A journey from CycleStreets.net}
\format{
An object of class \code{sf}
}
\source{
CycleStreets.net
}
\usage{
cyclestreets_route
}
\description{
Road segments representing suggested route to cycle in Leeds, UK.
}
\examples{
# cyclestreets_route = cyclestreets::journey(
#   stplanr::geo_code("Potternewton Park Leeds"),
#   stplanr::geo_code("University of Leeds")
# )
# usethis::use_data(cyclestreets_route)
class(cyclestreets_route)
# plot(cyclestreets_route)
}
\keyword{datasets}
