% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem_lisbon_raster}
\alias{dem_lisbon_raster}
\title{Elevation in central Lisbon, Portugal}
\format{
A raster dataset containing elevation above sea level
in a 1km bounding box in Lisbon, Portugal.
}
\source{
\url{https://github.com/rspatial/terra/issues/29}
}
\usage{
dem_lisbon_raster
}
\description{
A dataset containing elevation in and around Lisbon
with a geographic resolution of 10m.
}
\examples{
names(dem_lisbon_raster)
raster::plot(dem_lisbon_raster)
plot(lisbon_road_segments["Avg_Slope"], add = TRUE)
}
\keyword{datasets}
