% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{elevation_add}
\alias{elevation_add}
\title{Add elevations to linestring sf object}
\usage{
elevation_add(
  routes,
  dem = NULL,
  method = "bilinear",
  terra = has_terra() && methods::is(dem, "SpatRaster")
)
}
\arguments{
\item{routes}{sf object with LINESTRING geometries}

\item{dem}{RasterLayer or SpatRaster}

\item{method}{character; extraction method (default "bilinear")}

\item{terra}{logical; use terra if TRUE}
}
\description{
Add elevations to linestring sf object
}
