% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{elevation_extract}
\alias{elevation_extract}
\title{Extract elevations from coordinates}
\usage{
elevation_extract(
  m,
  e,
  method = "bilinear",
  terra = has_terra() && methods::is(e, "SpatRaster")
)
}
\arguments{
\item{m}{Matrix containing coordinates and elevations}

\item{e}{Raster overlapping with \code{r} and values representing elevations}

\item{method}{The method of estimating elevation at points,
passed to the \code{extract} function for extracting values from raster
datasets. Default: \code{"bilinear"}.}

\item{terra}{Should the \code{terra} package be used?
\code{TRUE} by default if the package is installed \emph{and}
if \code{e} is of class \code{SpatRast}}
}
\description{
Extract elevations from coordinates
}
\examples{
m = sf::st_coordinates(lisbon_road_segments[1, ])
e = dem_lisbon_raster
elevation_extract(m, e)
elevation_extract(m, e, method = "simple")
# uncomment the following lines to test with terra (experimental)
# u = paste0("https://github.com/ITSLeeds/slopes/",
#    "releases/download/0.0.0/dem_lisbon.tif" )
# if(!file.exists("dem_lisbon.tif")) download.file(u, "dem_lisbon.tif")
# et = terra::rast("dem_lisbon.tif")
# elevation_extract(m, et)
}
