% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_get.R
\name{elevation_get}
\alias{elevation_get}
\title{Get elevation data for routes}
\usage{
elevation_get(routes, ..., output_format = "raster")
}
\arguments{
\item{routes}{An sf object containing linestring geometries}

\item{...}{Additional arguments passed to ceramic::cc_elevation}

\item{output_format}{Character string specifying output format ("raster" or "terra")}
}
\value{
Elevation raster data covering the routes
}
\description{
Downloads elevation data using the ceramic package for given routes.
}
