% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_get.R
\name{elevation_get}
\alias{elevation_get}
\title{Get elevation data from hosted maptile services}
\usage{
elevation_get(routes, ...)
}
\arguments{
\item{routes}{Routes, the gradients of which are to be calculated.
The object must be of class \code{sf} with \code{LINESTRING} geometries.}

\item{...}{Options passed to \code{cc_elevation()}}
}
\description{
\code{elevation_get()} uses the \verb{[ceramic::cc_elevation()]} function to get
DEM data in raster format anywhere worldwide.
It requires an API that can be added by following guidance in the package's
\href{https://github.com/ITSLeeds/slopes#installation-for-dem-downloads}{README}.
}
\details{
Note: if you use the \code{cc_elevation()} function directly to get DEM data,
you can cache the data, as described in the package's
\href{https://github.com/hypertidy/ceramic#local-caching-of-tiles}{README}.
}
\examples{
# Time-consuming examples that require an internet connection and API key:
\donttest{
library(sf)
library(raster)
routes = cyclestreets_route
e = elevation_get(routes)
class(e)
e
plot(e)
plot(sf::st_geometry(routes), add = TRUE)
}
}
