% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_get.R
\name{elevation_get}
\alias{elevation_get}
\title{Get elevation data from hosted maptile services}
\usage{
elevation_get(routes, file = NULL, ...)
}
\arguments{
\item{routes}{Routes, the gradients of which are to be calculated.
The object must be of class \code{sf} with \code{LINESTRING} geometries.}

\item{file}{Where to save the resulting data if specified (not implemented)}

\item{...}{Options passed to \code{cc_elevation()}}
}
\description{
Get elevation data from hosted maptile services
}
\examples{
# Time-consuming examples that require an internet connectioin and API key:
\donttest{
routes = cyclestreets_route
e = elevation_get(routes)
class(e)
e
plot(e)
plot(sf::st_geometry(routes), add = TRUE)
}
}
