% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lisbon_road_segment}
\alias{lisbon_road_segment}
\alias{lisbon_road_segment_3d}
\alias{lisbon_road_segment_xyz_mapbox}
\title{A road segment in Lisbon, Portugal}
\format{
An object of class \code{sf}
}
\source{
Produced by ESRI's
3D Analyst extension
}
\usage{
lisbon_road_segment
}
\description{
A single road segment and a 3d version.
Different versions of this dataset are provided.
}
\details{
The \code{lisbon_road_segment} has 23 columns and 1 row.

The \code{lisbon_road_segment_xyz_mapbox} was created with:
\code{lisbon_road_segment_xyz_mapbox = elevation_add(lisbon_road_segment)}.
}
\examples{
lisbon_road_segment
lisbon_road_segment_3d
lisbon_road_segment_xyz_mapbox
}
\keyword{datasets}
