% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lisbon_road_segments}
\alias{lisbon_road_segments}
\title{Road segments in Lisbon}
\format{
And object of class \code{sf}
}
\source{
Produced by ESRI's \href{https://pro.arcgis.com/en/pro-app/help/analysis/3d-analyst/get-started-with-3d-analyst-in-pro.htm}{3D Analyst extension}
}
\usage{
lisbon_road_segments
}
\description{
A dataset representing road segments in Lisbon, with X, Y and Z (elevation) dimensions for each coordinate.
}
\examples{
names(lisbon_road_segments)
plot(lisbon_road_segments["Avg_Slope"])
}
\keyword{datasets}
