% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slope.R
\name{plot_dz}
\alias{plot_dz}
\title{Plot a digital elevation profile based on xyz data}
\usage{
plot_dz(
  d,
  z,
  fill = TRUE,
  p = ifelse(test = require(colorspace), colorspace::diverging_hcl,
    grDevices::terrain.colors)
)
}
\arguments{
\item{d}{Cumulative distance}

\item{z}{Elevations at points across a linestring}

\item{fill}{Should the profile be filled? \code{TRUE} by default.}

\item{p}{Color palette to use}
}
\description{
Plot a digital elevation profile based on xyz data
}
\examples{
r = lisbon_road_segment_3d
m = sf::st_coordinates(r)
d = cumsum(sequential_dist(m, lonlat = FALSE))
d = c(0, d)
z = m[, 3]
plot_dz(d, z)
}
