% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slope.R
\name{plot_dz}
\alias{plot_dz}
\title{Plot a digital elevation profile based on xyz data}
\usage{
plot_dz(
  d,
  z,
  fill = TRUE,
  horiz = FALSE,
  p = ifelse(test = requireNamespace("colorspace", quietly = TRUE),
    colorspace::diverging_hcl, grDevices::terrain.colors),
  ...,
  x = "top",
  col = "black",
  cex = 0.9,
  bg = grDevices::rgb(1, 1, 1, 0.8),
  title = "Slope colors (percentage gradient)",
  brks = c(3, 6, 10, 20, 40, 100),
  s = NULL,
  ncol = 4
)
}
\arguments{
\item{d}{Cumulative distance}

\item{z}{Elevations at points across a linestring}

\item{fill}{Should the profile be filled? \code{TRUE} by default}

\item{horiz}{Should the legend be horizontal (\code{FALSE} by default)}

\item{p}{Color palette to use}

\item{...}{Additional parameters to pass to legend}

\item{x}{Keyword, one of "bottomright", "bottom",
"bottomleft", "left", "topleft", "top", "topright", "right" and "center"}

\item{col}{Line colour}

\item{cex}{Legend size}

\item{bg}{Legend background colour}

\item{title}{Title of the legend}

\item{brks}{Breaks in colour palette to show.
\code{c(1, 3, 6, 10, 20, 40, 100)} by default.}

\item{s}{Sequence of numbers to show in legend}

\item{ncol}{Number of columns in legend}
}
\description{
Plot a digital elevation profile based on xyz data
}
\examples{
routes = lisbon_road_segment_3d
m = sf::st_coordinates(routes)
d = cumsum(sequential_dist(m, lonlat = FALSE))
d = c(0, d)
z = m[, 3]
# not exported
# plot_dz(d, z)
}
