% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slope.R
\name{plot_slope}
\alias{plot_slope}
\title{Plot slope data for a 3d linestring with base R graphics}
\usage{
plot_slope(r, fill = TRUE, lonlat = sf::st_is_longlat(r))
}
\arguments{
\item{r}{A linestring with xyz dimensions}

\item{fill}{Should the profile be filled? \code{TRUE} by default.}

\item{lonlat}{Are the coordinates in lon/lat order? \code{TRUE} by default}
}
\description{
Plot slope data for a 3d linestring with base R graphics
}
\examples{
plot_slope(lisbon_route_3d)
r = lisbon_road_segment_3d
plot_slope(r)
r = lisbon_road_segment_3d
}
