% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slope.R
\name{plot_slope}
\alias{plot_slope}
\title{Plot slope data for a 3d linestring with base R graphics}
\usage{
plot_slope(
  r,
  lonlat = sf::st_is_longlat(r),
  fill = TRUE,
  horiz = FALSE,
  p = ifelse(test = requireNamespace("colorspace", quietly = TRUE),
    colorspace::diverging_hcl, grDevices::terrain.colors),
  ...,
  x = "top",
  col = "black",
  cex = 0.9,
  bg = grDevices::rgb(1, 1, 1, 0.8),
  title = "Slope colors (percentage gradient)",
  s = 3:18,
  ncol = 4
)
}
\arguments{
\item{r}{A linestring with xyz dimensions}

\item{lonlat}{Are the coordinates in lon/lat order? \code{TRUE} by default}

\item{fill}{Should the profile be filled? \code{TRUE} by default}

\item{horiz}{Should the legend be horizontal (\code{FALSE} by default)}

\item{p}{Color palette to use}

\item{...}{Additional parameters to pass to legend}

\item{x}{Keyword, one of "bottomright", "bottom",
"bottomleft", "left", "topleft", "top", "topright", "right" and "center"}

\item{col}{Line colour}

\item{cex}{Legend size}

\item{bg}{Legend background colour}

\item{title}{Title of the legend}

\item{s}{Sequence of numbers to show in legend}

\item{ncol}{Number of columns in legend}
}
\description{
Plot slope data for a 3d linestring with base R graphics
}
\examples{
plot_slope(lisbon_route_3d)
r = lisbon_road_segment_3d
plot_slope(r)
r = lisbon_road_segment_3d
}
