% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{sequential_dist}
\alias{sequential_dist}
\title{Calculate sequential distances between points}
\usage{
sequential_dist(m, lonlat = TRUE)
}
\arguments{
\item{m}{Matrix of coordinates (x, y)}

\item{lonlat}{Logical, whether coordinates are longitude/latitude (default: TRUE)}
}
\value{
Numeric vector of distances between consecutive points
}
\description{
Calculates distances between consecutive points in a coordinate matrix.
}
