% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{sequential_dist}
\alias{sequential_dist}
\title{Calculate the sequential distances between sequential coordinate pairs}
\usage{
sequential_dist(m, lonlat = TRUE)
}
\arguments{
\item{m}{Matrix containing coordinates and elevations}

\item{lonlat}{Are the coordinates in lon/lat (geographic) coordinates? TRUE by default.}
}
\description{
Set \code{lonlat} to \code{FALSE} if you have projected data, e.g. with coordinates
representing distance in meters, not degrees. Lonlat coodinates are assumed
(\code{lonlat = TRUE} is the default).
}
\examples{
x = c(0, 2, 3, 4, 5, 9)
y = c(0, 0, 0, 0, 0, 1)
m = cbind(x, y)
sequential_dist(m)
}
