% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_3d}
\alias{slope_3d}
\title{Take a linestring and add a third (z) dimension to its coordinates}
\usage{
slope_3d(r, e = NULL, method = "bilinear", terra = has_terra())
}
\arguments{
\item{r}{Routes, the gradients of which are to be calculated.
The object must be of class \code{sf} with \code{LINESTRING} geometries.}

\item{e}{A raster object overlapping with \code{r} and values representing elevations}

\item{method}{The method of estimating elevation at points, passed to the \code{extract}
function for extracting values from raster datasets. Default: \code{"bilinear"}.}

\item{terra}{Should the \code{terra} package be used?
\code{TRUE} by default if the package is installed \emph{and}
if \code{e} is of class \code{SpatRast}}
}
\description{
Take a linestring and add a third (z) dimension to its coordinates
}
\examples{
r = lisbon_road_segments[204, ]
e = dem_lisbon_raster
(r3d = slope_3d(r, e))
sf::st_z_range(r)
sf::st_z_range(r3d)
plot(sf::st_coordinates(r3d)[, 3])
# (r3d = slope_3d(r, et))
# takes bandwidth and time and (currently) fails with:
#   GDAL Error 1: No PROJ.4 translation for source SRS
# https://github.com/ITSLeeds/slopes/runs/648128378#step:18:107
# r3d_get = slope_3d(cyclestreets_route)
# plot_slope(r3d_get)
}
