% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{slope_breaks}
\alias{slope_breaks}
\title{Recommended slope break thresholds}
\format{
A numeric vector of length 7.
}
\usage{
slope_breaks
}
\description{
A numeric vector of seven break points (as proportions, e.g. 0.05 = 5 \\%)
that define the slope severity classes used with \code{\link{slope_colors}}.
Classes are: 0--3\\%, 3--5\\%, 5--8\\%, 8--10\\%, 10--20\\%, >20\\%.
}
\examples{
slope_breaks
}
\seealso{
\code{\link{slope_colors}}, \code{\link{slopes_palette}}
}
