% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{slope_colors}
\alias{slope_colors}
\title{Recommended slope colours (dark green to dark red)}
\format{
A character vector of length 6.
}
\usage{
slope_colors
}
\description{
A character vector of six hex colours ranging from dark green (flat terrain)
to dark red (steep terrain), representing increasing slope severity.
Designed to be used together with \code{\link{slope_breaks}}.
}
\examples{
slope_colors
plot_slope(lisbon_route_xyz, pal = slope_colors, brks = slope_breaks)
}
\seealso{
\code{\link{slope_breaks}}, \code{\link{slopes_palette}}
}
