% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_distance}
\alias{slope_distance}
\title{Calculate distance-based slopes}
\usage{
slope_distance(d, elevations)
}
\arguments{
\item{d}{Distance values between consecutive points}
\item{elevations}{Elevation values at each point}
}
\value{
Numeric vector of slope gradients
}
\description{
Calculates slope gradients using distance between consecutive points in linestring geometries.
}
\details{
This function computes slope by considering the horizontal distance between vertices and the elevation change.
}
