% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_matrix}
\alias{slope_matrix}
\title{Calculate slopes from coordinate matrix}
\usage{
slope_matrix(m, elevations = m[, 3], lonlat = TRUE)
}
\arguments{
\item{m}{Matrix of coordinates (x, y, z)}

\item{elevations}{Vector of elevation values (default: third column of m)}

\item{lonlat}{Logical, whether coordinates are longitude/latitude (default: TRUE)}
}
\value{
Numeric vector of slope values
}
\description{
Calculates slope gradients from a matrix of coordinates and elevation data.
}
