% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_matrix_mean}
\alias{slope_matrix_mean}
\title{Calculate mean slope from coordinate matrix}
\usage{
slope_matrix_mean(m, elevations = m[, 3], lonlat = TRUE, directed = FALSE)
}
\arguments{
\item{m}{Matrix of coordinates (x, y, z)}

\item{elevations}{Vector of elevation values (default: third column of m)}

\item{lonlat}{Logical, whether coordinates are longitude/latitude (default: TRUE)}

\item{directed}{Logical, whether to calculate directed slopes (default: FALSE)}
}
\value{
Numeric value representing the mean slope
}
\description{
Computes the mean slope from a matrix of coordinates with elevation data.
}
