% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_get.R
\name{slope_matrix_to_raster}
\alias{slope_matrix_to_raster}
\title{Convert slope matrix to raster}
\usage{
slope_matrix_to_raster(x)
}
\arguments{
\item{x}{A matrix or RasterLayer object}
}
\value{
A RasterLayer object
}
\description{
Converts a slope matrix or checks if input is already a raster.
}
