% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_get.R, R/slopes.R
\name{slope_raster}
\alias{slope_raster}
\title{Convert a slope matrix to a raster object}
\usage{
slope_raster(
  routes,
  dem,
  lonlat = sf::st_is_longlat(routes),
  method = "bilinear",
  fun = slope_matrix_weighted,
  terra = has_terra() && methods::is(dem, "SpatRaster"),
  directed = FALSE
)

slope_raster(
  routes,
  dem,
  lonlat = sf::st_is_longlat(routes),
  method = "bilinear",
  fun = slope_matrix_weighted,
  terra = has_terra() && methods::is(dem, "SpatRaster"),
  directed = FALSE
)
}
\arguments{
\item{routes}{sf object with LINESTRING geometries}

\item{dem}{RasterLayer or SpatRaster}

\item{lonlat}{logical; are coordinates longitude/latitude? Default: sf::st_is_longlat(routes)}

\item{method}{character; method for extracting raster values (default "bilinear")}

\item{fun}{slope function to apply (default: slope_matrix_weighted)}

\item{terra}{logical; use terra if TRUE}

\item{directed}{logical; if TRUE, downslope is negative}

\item{x}{Matrix or RasterLayer of slope values}
}
\value{
A RasterLayer object
}
\description{
Convert a slope matrix to a raster object

Calculate slopes from raster and sf routes
}
