% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_raster}
\alias{slope_raster}
\title{Calculate the gradient of line segments from a raster dataset}
\usage{
slope_raster(
  r,
  e = NULL,
  lonlat = sf::st_is_longlat(r),
  method = "bilinear",
  fun = slope_matrix_weighted,
  terra = has_terra()
)
}
\arguments{
\item{r}{Routes, the gradients of which are to be calculated}

\item{e}{A raster object overlapping with \code{r} and values representing elevations}

\item{lonlat}{Are the coordinates in lon/lat order? \code{TRUE} by default}

\item{method}{The method of estimating elevation at points, passed to the \code{extract}
function for extracting values from raster datasets. Default: \code{"bilinear"}.}

\item{fun}{The slope function to calculate per route,
\code{slope_matrix_weighted} by default.}

\item{terra}{Should the \code{terra} package be used?
\code{TRUE} by default if the package is installed \emph{and}
if \code{e} is of class \code{SpatRast}}
}
\description{
Calculate the gradient of line segments from a raster dataset
}
\examples{
# r = lisbon_road_segments[239, ]
r = lisbon_road_segments
e = dem_lisbon_raster
(s = slope_raster(r, e))[1:5]
cor(r$Avg_Slope, s)
}
