% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_raster}
\alias{slope_raster}
\title{Calculate the gradient of line segments from a raster dataset}
\usage{
slope_raster(r, e = NULL, lonlat = FALSE, method = "bilinear")
}
\arguments{
\item{r}{Routes, the gradients of which are to be calculated}

\item{e}{A raster object overlapping with \code{r} and values representing elevations}

\item{lonlat}{Are the coordinates in lon/lat order? \code{TRUE} by default}

\item{method}{The method of estimating elevation at points, passed to the \code{extract}
function for extracting values from raster datasets. Default: \code{"bilinear"}.}
}
\description{
Calculate the gradient of line segments from a raster dataset
}
\examples{
r = lisbon_road_segments
e = dem_lisbon_raster
(s = slope_raster(r, e))
cor(r$Avg_Slope, s)^2
}
