% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_vector}
\alias{slope_vector}
\title{Calculate slopes from vector data}
\usage{
slope_vector(x, elevations)
}
\arguments{
\item{x}{Vector of distance values}

\item{elevations}{Vector of elevation values}
}
\value{
Numeric vector of slope values
}
\description{
Calculates slope gradients using vector distance and elevation data.
}
