% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_vector}
\alias{slope_vector}
\title{Calculate the gradient of line segments from distance and elevation vectors}
\usage{
slope_vector(x, e)
}
\arguments{
\item{x}{Vector of locations}

\item{e}{Elevations in same units as x (assumed to be metres)}
}
\description{
Calculate the gradient of line segments from distance and elevation vectors
}
\examples{
x = c(0, 2, 3, 4, 5, 9)
e = c(1, 2, 2, 4, 3, 1) / 10
slope_vector(x, e)
}
