% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_get.R, R/slopes.R
\name{slope_xyz}
\alias{slope_xyz}
\title{Extract slope values as xyz coordinates}
\usage{
slope_xyz(
  route_xyz,
  fun = slope_matrix_weighted,
  lonlat = TRUE,
  directed = FALSE
)

slope_xyz(
  route_xyz,
  fun = slope_matrix_weighted,
  lonlat = TRUE,
  directed = FALSE
)
}
\arguments{
\item{route_xyz}{data.frame or sf object}

\item{fun}{slope function (default: slope_matrix_weighted)}

\item{lonlat}{logical; are coordinates longitude/latitude?}

\item{directed}{logical; if TRUE, downslope is negative}

\item{x}{RasterLayer or matrix of slope values}
}
\value{
Data frame with x, y, z values
}
\description{
Extract slope values as xyz coordinates

Calculate slopes from xyz data frames or sf objects
}
