% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_xyz}
\alias{slope_xyz}
\title{Extract slopes from xyz data frame or sf objects}
\usage{
slope_xyz(
  route_xyz,
  fun = slope_matrix_weighted,
  lonlat = TRUE,
  directed = FALSE
)
}
\arguments{
\item{route_xyz}{An sf object with x, y, z dimensions}

\item{fun}{The slope function to calculate per route,
\code{slope_matrix_weighted} by default.}

\item{lonlat}{Are the coordinates in lon/lat order? TRUE by default}

\item{directed}{Should the value be directed? \code{FALSE} by default.
If \code{TRUE} the result will be negative when it represents a downslope
(when the end point is lower than the start point).}
}
\value{
A vector of slopes equal in length to the number simple features
(rows representing linestrings) in the input object.
}
\description{
Extract slopes from xyz data frame or sf objects
}
\examples{
route_xyz = lisbon_road_segment_3d
slope_xyz(route_xyz, lonlat = FALSE)
slope_xyz(route_xyz, lonlat = FALSE, directed = TRUE)
}
