% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{slope_xyz}
\alias{slope_xyz}
\title{Extract slopes from xyz dataframe or sf objects}
\usage{
slope_xyz(r_xyz, fun = slope_matrix_weighted, lonlat = NULL)
}
\arguments{
\item{r_xyz}{An sf object with x, y, z dimensions}

\item{fun}{The slope function to calculate per route,
\code{slope_matrix_weighted} by default.}

\item{lonlat}{Are the coordinates in lon/lat order? \code{TRUE} by default}
}
\description{
Extract slopes from xyz dataframe or sf objects
}
\examples{
r_xyz = lisbon_road_segment_3d
slope_xyz(r_xyz)
}
