% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_get.R
\name{slope_xyz_simple}
\alias{slope_xyz_simple}
\title{Extract slope values as xyz coordinates}
\usage{
slope_xyz_simple(x)
}
\arguments{
\item{x}{RasterLayer or matrix of slope values}
}
\value{
Data frame with x, y, z values
}
\description{
Extract slope values as xyz coordinates
}
