% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes-package.R
\docType{package}
\name{slopes-package}
\alias{slopes}
\alias{slopes-package}
\title{slopes: Calculate Slopes of Roads, Rivers and Trajectories}
\description{
Calculate the slope (also known as steepness and gradient)
of linear features such as roads, rivers and ski routes.

Created by Robin Lovelace and Rosa Félix, who developed the package
to support their research in active travel, and the prioritisation of
investment in cycling, an activity which is sensitive to slopes,
in particular.

See \href{https://github.com/ITSLeeds/slopes/}{github.com/ITSLeeds/slopes}
for the source code and \href{https://ropensci.github.io/slopes/}{ropensci.github.io/slopes}
for the website.

All package functions can be found in the
\href{https://ropensci.github.io/slopes/reference/index.html}{Reference page}
on the package website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/slopes/}
  \item \url{https://ropensci.github.io/slopes/}
  \item Report bugs at \url{https://github.com/ropensci/slopes/issues}
}

}
\author{
\strong{Maintainer}: Robin Lovelace \email{rob00x@gmail.com} (\href{https://orcid.org/0000-0001-5679-6536}{ORCID})

Authors:
\itemize{
  \item Joey Talbot \email{j.d.talbot@leeds.ac.uk} (\href{https://orcid.org/0000-0002-6520-4560}{ORCID})
  \item Rosa Félix \email{rosamfelix@tecnico.ulisboa.pt} (\href{https://orcid.org/0000-0002-5642-6006}{ORCID})
}

}
\keyword{internal}
