% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{slopes_palette}
\alias{slopes_palette}
\title{Get a slope color palette}
\usage{
slopes_palette(n = 6, palette = "Green-Brown")
}
\arguments{
\item{n}{Number of colors (default 6)}

\item{palette}{Name of palette, default "Green-Brown"}
}
\value{
A vector of hex color codes
}
\description{
Get a slope color palette
}
