% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.R
\name{z_value}
\alias{z_value}
\alias{z_start}
\alias{z_end}
\alias{z_mean}
\alias{z_max}
\alias{z_min}
\title{Calculate summary values for 'Z' elevation attributes}
\usage{
z_value(x)

z_start(x)

z_end(x)

z_mean(x)

z_max(x)

z_min(x)
}
\arguments{
\item{x}{An \code{sfc} object with 'XYZ' coordinates}
}
\description{
The \verb{slope_z*()} functions calculate summary values for the Z axis
in \code{sfc} objects with \code{XYZ} geometries.
}
\examples{
x = slopes::lisbon_route_3d
sf::st_geometry(x)
z_value(x)[1:5]
xy = slopes::lisbon_route
try(z_value(xy)) # error message
z_start(x)
z_end(x)
}
